Update and rename FaceBook.php to PhotoBook.php
I've change the name of the file.
